requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
python-dotenv==0.18.0
requests==2.25.1
pytest==6.2.4